[App] Show published status of books
