+++ peak-app/src/common/notes/list/note-item-metadata-container/metadata-container.scss
@@ -1,0 +1,21 @@
+@import "../../../../constants/variables";
+
+.metadata-container {
+  display: flex;
+  align-items: center;
+
+  .publish-tag {
+    border-color: $primary-green;
+    color: $primary-green;
+    margin-right: 10px;
+  }
+
+  .space-holder {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+
+    width: 20px;
+    height: 20px;
+  }
+}
