+++ peak-app/src/views/timeline/peak-timeline.scss
@@ -102,32 +102,11 @@
             align-items: center;
             justify-content: space-between;
 
-            .metadata-container {
-              display: flex;
-              align-items: center;
-
-              .publish-tag {
-                border-color: $primary-green;
-                color: $primary-green;
-                margin-right: 10px;
-              }
-
-              .space-holder {
-                display: flex;
-                justify-content: center;
-                align-items: center;
-
-                width: 20px;
-                height: 20px;
-              }
-
-
-            }
-
             .peak-timeline-item-body {
               display: flex;
               flex-direction: column;
               flex-grow: 1;
+              padding-right: 25px;
 
               .title {
                 font-size: 1.25rem;
